+++ CHANGELOG.md
@@ -17,6 +17,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Beta User Acceptance Testing]
 - Interaction Monitoring
 
+## [0.36.0] - 2018-11-05
+### Added
+- Users page.. ability to add and remove 
+  outbound identifier lists, 
+  message templates, 
+  skills,
+  groups,
+  transfer lists,
+  presence reasons
+
 ## [0.35.0] - 2018-11-05
 ### Added
 - Skills and groups members lists.. ability to add and remove outbound identifier lists